src/main/kotlin/com/tomassirio/easyinstaller/service/impl/installer/RipGrepInstaller.kt
package com.tomassirio.easyinstaller.service.impl.installer

import com.tomassirio.easyinstaller.service.InstallableApplication
import com.tomassirio.easyinstaller.service.annotation.CommandLineTool
import com.tomassirio.easyinstaller.service.impl.installer.builder.DefaultCommandBuilder
import com.tomassirio.easyinstaller.service.impl.installer.strategy.DownloadStrategyContext
import com.tomassirio.easyinstaller.style.ShellFormatter
import org.springframework.beans.factory.annotation.Value
import org.springframework.stereotype.Service

@Service
@CommandLineTool
class RipGrepInstaller(
    private val shellFormatter: ShellFormatter,
    private val downloadStrategyContext: DownloadStrategyContext
): InstallableApplication {

    @Value("\${url.default.ripgrep}")
    lateinit var DEFAULT_URL: String

    override fun install() {
        shellFormatter.printInfo("Installing ${name()}...")
        val strategy = downloadStrategyContext.getCurrentStrategy()
        val command = if (downloadStrategyContext.isDefault()) createDefaultCommand() else name().lowercase()
        strategy(command)
    }

    override fun name() = "RipGrep"

    private fun createDefaultCommand(): String {
        return DefaultCommandBuilder(name(), DEFAULT_URL)
                .setExtractCommand("tar -xz")
                .addPostExtractCommands(
                        "cd ripgrep-14.0.3-x86_64-unknown-linux-musl",
                        "sudo cp rg /usr/local/bin/")
                .addCleanupCommands(
                        "rm -rf ripgrep-14.0.3-x86_64-unknown-linux-musl")
                .useSudo()
                .build()
    }
}